deploy v0.6.6
This version introduces a new -- and still evolving -- feature called "decompilation".

Decompilation allows you to recover the blocks source code of a MicroBlocks program stored on a board. The decompiler is not yet able to recover some things, such as comments, local variable and parameter names, and some function names, but the decompiled code will run and can be edited. The decompiler will evolve over time to recover more of the original names and other details.

If you'd like to to try the decompiler in its current form, first select "show advanced blocks" in the MicroBlocks (gear) menu. Then select "enable fetching code from the board on connection".

Before you plug in the board whose code you want to decompile, first select "New" to clear the current project. Then plug in the board. The scripts stored on the board will appear as if by magic!

This feature is still under development, which is why it is disabled by default. But it's so cool we wanted to share it those interested in trying it. :-)

System
* first version of the decompiler (can recreate blocks from the code stored in the board)
* add a preference to enable/disable decompiler (disabled by default)
* remove the "download and install latest VM" advanced menu command (since the latest installable VM's are built into the IDE)
* upgraded to PlatformIO version 5.0.0 for compiling VM's

Bug Fixes
* Fix: Ensure IP address is 0.0.0.0 when not connected
* Fix: Don't raise error when "find" called on a non-indexable object, just return -1 (not found)
* Fix: Incorrect default NeoPixel pin on Clue